requirements.txt
Flask==2.3.0
discord.py==2.7.0
